ConductanceSystem

A ConductanceSystem describes the state- and time-dependence of a conductance (i.e. classically g, the inverse of resistance; typically measured in Siemens). It can be used to model the conductance associated with ionic membrane currents, synaptic currents, and axial currents that flow between connected neuronal compartments. By default, a Conductance System is composed of zero or more gates and a parameter for the maximum conductance value ($\overline{g}$). The output of a ConductanceSystem is equal to the the product of each individual gate and $\overline{g}$.

To model sodium conductance in a Hodgkin-Huxley model:

using Conductor, ModelingToolkit, IfElse, Unitful
import Unitful: mV, mS, cm

Vₘ = MembranePotential()

nav_kinetics = [
    Gate(AlphaBeta,
         IfElse.ifelse(Vₘ == -40.0, 1.0, (0.1*(Vₘ + 40.0))/(1.0 - exp(-(Vₘ + 40.0)/10.0))),
         4.0*exp(-(Vₘ + 65.0)/18.0), p = 3, name = :m)
    Gate(AlphaBeta,
         0.07*exp(-(Vₘ+65.0)/20.0),
         1.0/(1.0 + exp(-(Vₘ + 35.0)/10.0)), name = :h)]

@named NaV = IonChannel(Sodium, nav_kinetics, max_g = 120mS/cm^2)

equations(NaV) # includes: g(t) ~ gbar*(m(t)^3)*h(t)

\begin{align} \frac{dm(t)}{dt} =& \left( 1 - m\left( t \right) \right) \mathrm{ifelse}\left( 1.0, \frac{4.0 + 0.1 \mathrm{Vm}\left( t \right)}{1.0 - e^{-4.0 - 0.1 \mathrm{Vm}\left( t \right)}}; \left( \mathrm{Vm}\left( t \right) = -40.0 \right) \right) - 4.0 m\left( t \right) e^{-3.6111111111111107 - 0.05555555555555555 \mathrm{Vm}\left( t \right)} \ \frac{dh(t)}{dt} =& \frac{ - h\left( t \right)}{1.0 + e^{-3.5 - 0.1 \mathrm{Vm}\left( t \right)}} + 0.07 \left( 1 - h\left( t \right) \right) e^{-3.25 - 0.05 \mathrm{Vm}\left( t \right)} \ g\left( t \right) =& \left( m\left( t \right) \right)^{3} gbar h\left( t \right) \end{align}

Conductor.ConductanceSystem — Type

struct ConductanceSystem <: Conductor.AbstractConductanceSystem

A model of conductance.

eqs

iv

Independent variabe. Defaults to time, $t$.

states

ps

observed

name

systems

defaults

output

Conductance, $g$, of the system.

gbar

Maximum conductance, $\overline{g}$.

ion

Permeability of the conductance.

aggregate

gate_vars

Gating variables.

subscriptions

Extrinsic sources of state (e.g. presynaptic compartments).

extensions

Additional systems to extend dynamics. Extensions are composed with the parent system during conversion to ODESystem.

inputs

Conductor.IonChannel — Function

IonChannel(ion, gate_vars; <keyword arguments>)

An ionic membrane conductance.

Arguments

max_g: Default value for maximum conductance, $\overline{g}$.

extensions::Vector{ODESystem}: Additional systems to extend dynamics. Extensions are composed with the parent system during conversion to ODESystem.

defaults::Dict: Default values for states and parameters.

name::Symbol: Name of the system.

Conductor.AxialConductance — Function

AxialConductance(gate_vars; <keyword arguments>)

A non-specific conductance between morphologically contiguous compartments.

Arguments

max_g: Maximum conductance, $\overline{g}$.

extensions::Vector{ODESystem}: Additional systems to extend dynamics. Extensions are composed with the parent system during conversion to ODESystem.

defaults::Dict: Default values for states and parameters.

name::Symbol: Name of the system.

Conductor.SynapticChannel — Function

SynapticChannel(ion, gate_vars; <keyword arguments>)

A synaptically activated conductance. Depends on extrinsic (i.e. presynaptic) state.

Arguments

max_g: Maximum conductance, $\overline{g}$.

extensions::Vector{ODESystem}: Additional systems to extend dynamics. Extensions are composed with the parent system during conversion to ODESystem.

aggregate::Bool: whether the Conductance model should aggregate extrinsic sources of state instead of integrating them independently. Defaults to false.

defaults::Dict: Default values for states and parameters.

name::Symbol: Name of the system.