Networks

NeuronalNetworkSystem

Conductor.Population — Type

struct Population{T<:Conductor.AbstractCompartmentSystem}

An iterable population (or small group) of neurons.

A Population is a lazy representation of neurons templated from a single model prototype. Iterating over a Population returns up to n uniquely-named copies of the prototype that are serially numbered and, optionally, namespaced. Parameter values of the replicate neurons can be individually defined by indexed assignment, or applied to the entire group via distributions.

proto

The base neuron model.

n

The number of neurons in the population.

popname

Name of the population.

defs

Population defaults

stimuli

Stimuli applied to select neurons.

Conductor.NeuronalNetworkSystem — Type

struct NeuronalNetworkSystem <: Conductor.AbstractNeuronalNetworkSystem

A network of neurons with synaptic connections.

eqs::Vector{Symbolics.Equation}

iv::Symbolics.Num

Independent variabe. Defaults to time, $t$.

states::Vector{Symbolics.Num}

ps::Vector{Symbolics.Num}

observed::Vector{Symbolics.Equation}

name::Symbol

systems::Vector{ModelingToolkit.AbstractTimeDependentSystem}

defaults::Dict

topology::NetworkTopology

reversal_map::Dict

extensions::Vector{ModelingToolkit.ODESystem}

Additional systems to extend dynamics. Extensions are composed with the parent system during conversion to ODESystem.