Compartments

Single Compartments

A CompartmentSystem is sufficient to model a simple neuron–either as an isopotential spherical neuron, or as a "point" neuron with no morphological geometry. At minimum, a CompartmentSystem will contain information about its instrinsic membrane currents and ionic gradients (i.e. equilbrium potentials). We define these by directly providing vectors of ConductanceSystem and EquilibriumPotential to the CompartmentSystem:

using Conductor, Unitful, ModelingToolkit, IfElse
import Unitful: mV, mS, cm

Vₘ = MembranePotential()

nav_kinetics = [
    Gate(AlphaBeta,
         IfElse.ifelse(Vₘ == -40.0, 1.0, (0.1*(Vₘ + 40.0))/(1.0 - exp(-(Vₘ + 40.0)/10.0))),
         4.0*exp(-(Vₘ + 65.0)/18.0), p = 3, name = :m)
    Gate(AlphaBeta,
         0.07*exp(-(Vₘ+65.0)/20.0),
         1.0/(1.0 + exp(-(Vₘ + 35.0)/10.0)), name = :h)]

kdr_kinetics = [
    Gate(AlphaBeta,
         IfElse.ifelse(Vₘ == -55.0, 0.1, (0.01*(Vₘ + 55.0))/(1.0 - exp(-(Vₘ + 55.0)/10.0))),
         0.125 * exp(-(Vₘ + 65.0)/80.0),
         p = 4, name = :n)]

# Note: `IonChannel` is a convenience constructor that returns a `ConductanceSystem`
@named NaV = IonChannel(Sodium, nav_kinetics, max_g = 120mS/cm^2)
@named Kdr = IonChannel(Potassium, kdr_kinetics, max_g = 36mS/cm^2)
@named leak = IonChannel(Leak, max_g = 0.3mS/cm^2)

channels = [NaV, Kdr, leak];
reversals = Equilibria([Sodium => 50.0mV, Potassium => -77.0mV, Leak => -54.4mV])

@named neuron = CompartmentSystem(Vₘ, channels, reversals)

\[ \begin{align} \mathrm{INaV}\left( t \right) =& a_m \left( - ENa + \mathrm{V_m}\left( t \right) \right) \mathrm{NaV_{+}g}\left( t \right) \\ \mathrm{IKdr}\left( t \right) =& a_m \left( - EK + \mathrm{V_m}\left( t \right) \right) \mathrm{Kdr_{+}g}\left( t \right) \\ \mathrm{Ileak}\left( t \right) =& a_m \left( - El + \mathrm{V_m}\left( t \right) \right) \mathrm{leak_{+}g}\left( t \right) \\ \frac{dV_m(t)}{dt} =& \frac{ - \mathrm{IKdr}\left( t \right) - \mathrm{INaV}\left( t \right) - \mathrm{Ileak}\left( t \right)}{a_m c_m} \\ \mathrm{V_m}\left( t \right) =& \mathrm{NaV_{+}V_m}\left( t \right) \\ \mathrm{V_m}\left( t \right) =& \mathrm{Kdr_{+}V_m}\left( t \right) \\ \frac{dNaV_{+}m(t)}{dt} =& \left( 1 - \mathrm{NaV_{+}m}\left( t \right) \right) \mathrm{ifelse}\left( 1.0, \frac{4.0 + 0.1 \mathrm{NaV_{+}V_m}\left( t \right)}{1.0 - e^{-4.0 - 0.1 \mathrm{NaV_{+}V_m}\left( t \right)}}; \left( \mathrm{NaV_{+}V_m}\left( t \right) = -40.0 \right) \right) - 4.0 \mathrm{NaV_{+}m}\left( t \right) e^{-3.6111111111111107 - 0.05555555555555555 \mathrm{NaV_{+}V_m}\left( t \right)} \\ \frac{dNaV_{+}h(t)}{dt} =& \frac{ - \mathrm{NaV_{+}h}\left( t \right)}{1.0 + e^{-3.5 - 0.1 \mathrm{NaV_{+}V_m}\left( t \right)}} + 0.07 \left( 1 - \mathrm{NaV_{+}h}\left( t \right) \right) e^{-3.25 - 0.05 \mathrm{NaV_{+}V_m}\left( t \right)} \\ \mathrm{NaV_{+}g}\left( t \right) =& \left( \mathrm{NaV_{+}m}\left( t \right) \right)^{3} NaV_{+}gbar \mathrm{NaV_{+}h}\left( t \right) \\ \frac{dKdr_{+}n(t)}{dt} =& \left( 1 - \mathrm{Kdr_{+}n}\left( t \right) \right) \mathrm{ifelse}\left( 0.1, \frac{0.55 + 0.01 \mathrm{Kdr_{+}V_m}\left( t \right)}{1.0 - e^{-5.5 - 0.1 \mathrm{Kdr_{+}V_m}\left( t \right)}}; \left( \mathrm{Kdr_{+}V_m}\left( t \right) = -55.0 \right) \right) - 0.125 \mathrm{Kdr_{+}n}\left( t \right) e^{-0.8125 - 0.0125 \mathrm{Kdr_{+}V_m}\left( t \right)} \\ \mathrm{Kdr_{+}g}\left( t \right) =& \left( \mathrm{Kdr_{+}n}\left( t \right) \right)^{4} Kdr_{+}gbar \\ \mathrm{leak_{+}g}\left( t \right) =& leak_{+}gbar \end{align} \]

Note

We use IfElse.ifelse to handle discontinuities in the rate equations for Sodium and Potassium channel dynamics. In Julia v1.8+ Base.ifelse may be used directly without the need for an external dependency.

In the case above, the CompartmentSystem constructor assumes a dimensionless geometry = Point(). The maximum magnitudes of the ion channel conductances, $\overline{g}$, have units of SpecificConductance (mS/cm²) that scale with the surface area of the compartment. However, when the geometry is a Point(), the CompartmentSystem ignores surface area scaling (internally, the area is fixed to 1.0). We can instead provide a <:Geometry object to describe the shape and size of the compartment:

import Unitful: µm
soma_shape = Sphere(radius = 20µm)

@named neuron = CompartmentSystem(Vₘ, channels, reversals; geometry = soma_shape)

If we try to simulate the neuron we've modeled so far, the result isn't very interesting:

import Unitful: ms
using OrdinaryDiffEq, Plots
sim = Simulation(neuron, time = 300ms)
solution = solve(sim, Rosenbrock23())
plot(solution; vars=[Vₘ])

The neuron isn't spontaneously active. To make the neuron produce spikes, we can write an equation for an electrode current and provide it to CompartmentSystem:

import Unitful: µA, pA
@named Iₑ = IonCurrent(NonIonic)

# A 400 picoamp squarewave pulse when 100ms > t > 200ms
electrode_pulse = Iₑ ~ IfElse.ifelse(t > 100.0,
                                     IfElse.ifelse(t < 200.0,
                                                   ustrip(Float64, µA, 400pA),
                                                   0.0),
                                     0.0)

@named neuron_stim = CompartmentSystem(Vₘ, channels, reversals;
                                       geometry = soma_shape,
                                       stimuli = [electrode_pulse])

\[ \begin{align} \mathrm{INaV}\left( t \right) =& a_m \left( - ENa + \mathrm{V_m}\left( t \right) \right) \mathrm{NaV_{+}g}\left( t \right) \\ \mathrm{IKdr}\left( t \right) =& a_m \left( - EK + \mathrm{V_m}\left( t \right) \right) \mathrm{Kdr_{+}g}\left( t \right) \\ \mathrm{Ileak}\left( t \right) =& a_m \left( - El + \mathrm{V_m}\left( t \right) \right) \mathrm{leak_{+}g}\left( t \right) \\ \mathrm{I_e}\left( t \right) =& \mathrm{ifelse}\left( \left( t > 100.0 \right), \mathrm{ifelse}\left( \left( t < 200.0 \right), 0.0004, 0.0 \right), 0.0 \right) \\ \frac{dV_m(t)}{dt} =& \frac{ - \mathrm{IKdr}\left( t \right) - \mathrm{INaV}\left( t \right) - \mathrm{Ileak}\left( t \right) + \mathrm{I_e}\left( t \right)}{a_m c_m} \\ \mathrm{V_m}\left( t \right) =& \mathrm{NaV_{+}V_m}\left( t \right) \\ \mathrm{V_m}\left( t \right) =& \mathrm{Kdr_{+}V_m}\left( t \right) \\ \frac{dNaV_{+}m(t)}{dt} =& \left( 1 - \mathrm{NaV_{+}m}\left( t \right) \right) \mathrm{ifelse}\left( 1.0, \frac{4.0 + 0.1 \mathrm{NaV_{+}V_m}\left( t \right)}{1.0 - e^{-4.0 - 0.1 \mathrm{NaV_{+}V_m}\left( t \right)}}; \left( \mathrm{NaV_{+}V_m}\left( t \right) = -40.0 \right) \right) - 4.0 \mathrm{NaV_{+}m}\left( t \right) e^{-3.6111111111111107 - 0.05555555555555555 \mathrm{NaV_{+}V_m}\left( t \right)} \\ \frac{dNaV_{+}h(t)}{dt} =& \frac{ - \mathrm{NaV_{+}h}\left( t \right)}{1.0 + e^{-3.5 - 0.1 \mathrm{NaV_{+}V_m}\left( t \right)}} + 0.07 \left( 1 - \mathrm{NaV_{+}h}\left( t \right) \right) e^{-3.25 - 0.05 \mathrm{NaV_{+}V_m}\left( t \right)} \\ \mathrm{NaV_{+}g}\left( t \right) =& \left( \mathrm{NaV_{+}m}\left( t \right) \right)^{3} NaV_{+}gbar \mathrm{NaV_{+}h}\left( t \right) \\ \frac{dKdr_{+}n(t)}{dt} =& \left( 1 - \mathrm{Kdr_{+}n}\left( t \right) \right) \mathrm{ifelse}\left( 0.1, \frac{0.55 + 0.01 \mathrm{Kdr_{+}V_m}\left( t \right)}{1.0 - e^{-5.5 - 0.1 \mathrm{Kdr_{+}V_m}\left( t \right)}}; \left( \mathrm{Kdr_{+}V_m}\left( t \right) = -55.0 \right) \right) - 0.125 \mathrm{Kdr_{+}n}\left( t \right) e^{-0.8125 - 0.0125 \mathrm{Kdr_{+}V_m}\left( t \right)} \\ \mathrm{Kdr_{+}g}\left( t \right) =& \left( \mathrm{Kdr_{+}n}\left( t \right) \right)^{4} Kdr_{+}gbar \\ \mathrm{leak_{+}g}\left( t \right) =& leak_{+}gbar \end{align} \]

Putting it all together, our neuron simulation now produces a train of action potentials:

sim = Simulation(neuron_stim, time = 300ms)
solution = solve(sim, Rosenbrock23())
plot(solution; vars=[Vₘ])

Multiple Compartments

Neurons with multiple compartments are explicitly constructed by defining a MultiCompartmentTopology, a directed graph that is provided to MultiCompartmentSystem. Individual subcompartments are connected via add_junction!. By default, a junction between compartments is assumed to be symmetrical–the axial conductance from compartment A to B is the same value as from compartment B to A. However, if the conductance between compartments is asymmetric, two junctions must be defined–one for the conductance in each direction.

For example usage, see the pinskyrinzel.jl demo.

Conductor.CompartmentSystem — Type

struct CompartmentSystem <: Conductor.AbstractCompartmentSystem

A neuronal compartment.

eqs::Vector{Symbolics.Equation}

iv::Symbolics.Num

Independent variabe. Defaults to time, $t$.

states::Vector{Symbolics.Num}

ps::Vector{Symbolics.Num}

observed::Vector{Symbolics.Equation}

name::Symbol

systems::Vector{ModelingToolkit.AbstractTimeDependentSystem}

defaults::Dict

voltage::Symbolics.Num

Voltage potential.

capacitance::Symbolics.Num

Membrane capacitance.

geometry::Conductor.Geometry

Morphological geometry of the compartment.

chans::Vector{Conductor.AbstractConductanceSystem}

Ionic conductances.

channel_reversals::Vector{Symbolics.Num}

Equilibrium potentials belonging to ionic membrane conductances.

synapses::Vector{Conductor.AbstractConductanceSystem}

Synaptic conductances.

synaptic_reversals::Vector{Symbolics.Num}

Equilibrium potentials belonging to synaptic conductances.

axial_conductance::Vector{Tuple{Conductor.AbstractConductanceSystem, Symbolics.Num}}

Axial (intercompartmental) conductances.

stimuli::Vector{Symbolics.Equation}

Experimental stimuli (for example, current injection).

extensions::Vector{ModelingToolkit.ODESystem}

Additional systems to extend dynamics. Extensions are composed with the parent system during conversion to ODESystem.

parent::Ref{Conductor.AbstractCompartmentSystem}

Refers to the parent system when the compartment is a subcompartment in a MultiCompartmentSystem.

Conductor.MultiCompartmentSystem — Type

struct MultiCompartmentSystem <: Conductor.AbstractCompartmentSystem

A neuron with 2+ morphologically connected compartments.

eqs::Vector{Symbolics.Equation}

iv::Symbolics.Num

Independent variabe. Defaults to time, $t$.

states::Vector{Symbolics.Num}

ps::Vector{Symbolics.Num}

observed::Vector{Symbolics.Equation}

name::Symbol

systems::Vector{ModelingToolkit.AbstractTimeDependentSystem}

defaults::Dict

topology::MultiCompartmentTopology

compartments::Vector{CompartmentSystem}

Individual subcompartments of the neuron.

extensions::Vector{ModelingToolkit.ODESystem}

Additional systems to extend dynamics. Extensions are composed with the parent system during conversion to ODESystem.

Conductor.MultiCompartment — Type

MultiCompartment(topology; extensions, name, defaults) -> MultiCompartmentSystem

Basic constructor for a MultiCompartmentSystem.